+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Integration tests workflow
+* Integration tests via GHA (non-user endpoints)
 
 ## [2.21.0] - 2022-09-26